Implemented basic MovingConstantLogic

+++ ecs/src/main/java/space/battle/entity/component/system/dataclasses/Position.java
@@ -1,27 +1,27 @@
 package space.battle.entity.component.system.dataclasses;
 
 public class Position {
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 
 	public Position (int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX () {
+	public float getX () {
 		return x;
 	}
 
-	public void setX (int x) {
+	public void setX (float x) {
 		this.x = x;
 	}
 
-	public int getY () {
+	public float getY () {
 		return y;
 	}
 
-	public void setY (int y) {
+	public void setY (float y) {
 		this.y = y;
 	}
 }
